+++ src/2020/buri/aoc/y20/d21/Day21Test.java
@@ -27,7 +27,7 @@ public class Day21Test extends BaseTest {
 	 */
 	@Test
 	public void testPart1Puzzle() {
-		int result = Day21.getResult(Part.ONE, Day21.getInput(0));
+		long result = Day21.getResult(Part.ONE, Day21.getInput(0));
 		toConsole(result);
 		assertEquals(0, result);
 	}
@@ -42,7 +42,7 @@ public class Day21Test extends BaseTest {
 	 */
 	@Test
 	public void testPart2Puzzle() {
-		int result = Day21.getResult(Part.TWO, Day21.getInput(0));
+		long result = Day21.getResult(Part.TWO, Day21.getInput(0));
 		toConsole(result);
 		assertEquals(0, result);
 	}
